Remove unnecessary `requests` import failing tests
If `requests` becomes necessary in the future, it must be added to one
of the requirements lists in `setup.py`, otherwise Jenkins will fail to
run them (and so will anyone who has not got it installed already).

I don't think it will become necessary, because Flask provides HTTP
querying capabilities.