Improve error handling and reporting
Kinds of exceptions which cause InternalServerErrors are now caught, and
render a meaningful error page. If the user is an admin, they are given
a full stack trace and intructions to report the issue on the
repository. If the user is not an admin, they are given an incident
number to inform an admin about, while the stack trace is printed in the
logs.

This deals with the hard part of #49, but it is still necessary to find
places were we can prevent unexpected errors, instead of falling back
gracefully when they occur.